Client: Split WebRTC Service into smaller services.
- Introduced a new services for handling WebRTC communication and for handling the WebRTC Signaling.
